+++ src/main/java/roomescape/repository/MemberDao.java
@@ -1,40 +1,12 @@
 package roomescape.repository;
 
 import java.util.List;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Repository;
 import roomescape.domain.Member;
-import roomescape.domain.Role;
 
-@Repository
-public class MemberDao {
+public interface MemberDao {
+    List<Member> findAll();
 
-    private final JdbcTemplate jdbcTemplate;
-    private final RowMapper<Member> memberRowMapper = (resultSet, rowNum) -> new Member(
-            resultSet.getLong("id"),
-            resultSet.getString("name"),
-            resultSet.getString("email"),
-            resultSet.getString("password"),
-            Role.from(resultSet.getString("role"))
-    );
+    Member findByEmail(String email);
 
-    public MemberDao(final JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public List<Member> findAll() {
-        String sql = "SELECT * FROM member";
-        return jdbcTemplate.query(sql,memberRowMapper);
-    }
-
-    public Member findByEmail(final String email) {
-        String sql = "SELECT * FROM member WHERE email = ?";
-        return jdbcTemplate.queryForObject(sql, memberRowMapper, email);
-    }
-
-    public Member findById(final Long id) {
-        String sql = "SELECT * FROM member WHERE id = ?";
-        return jdbcTemplate.queryForObject(sql, memberRowMapper, id);
-    }
+    Member findById(Long id);
 }
