+++ src/main/java/roomescape/repository/repositoryImpl/JdbcMemberDao.java
@@ -1,0 +1,44 @@
+package roomescape.repository.repositoryImpl;
+
+import java.util.List;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Repository;
+import roomescape.domain.Member;
+import roomescape.domain.Role;
+import roomescape.repository.MemberDao;
+
+@Repository
+public class JdbcMemberDao implements MemberDao {
+
+    private final JdbcTemplate jdbcTemplate;
+    private final RowMapper<Member> memberRowMapper = (resultSet, rowNum) -> new Member(
+            resultSet.getLong("id"),
+            resultSet.getString("name"),
+            resultSet.getString("email"),
+            resultSet.getString("password"),
+            Role.from(resultSet.getString("role"))
+    );
+
+    public JdbcMemberDao(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public List<Member> findAll() {
+        String sql = "SELECT * FROM member";
+        return jdbcTemplate.query(sql, memberRowMapper);
+    }
+
+    @Override
+    public Member findByEmail(final String email) {
+        String sql = "SELECT * FROM member WHERE email = ?";
+        return jdbcTemplate.queryForObject(sql, memberRowMapper, email);
+    }
+
+    @Override
+    public Member findById(final Long id) {
+        String sql = "SELECT * FROM member WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, memberRowMapper, id);
+    }
+}
